Some improvements to let admins register new users.
refs #1361
