+++ app/views/users/new_user.html.haml
@@ -1,5 +1,5 @@
 #user-new
   .modal-header
     = modal_close_button
-    %h3= t('.title')
+    %h3= t('admins.new_user.title')
   = render 'form', :button_name => t('_other.create')
